requirements.txt
absl-py==1.4.0
accelerate==0.19.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==23.1.0
blobfile==2.0.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
einops==0.6.1
einops-exts==0.0.4
ema-pytorch==0.2.3
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
h5py==3.8.0
hickle==5.0.2
huggingface-hub==0.14.1
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
kiwisolver==1.4.4
lightning-utilities==0.8.0
lpips==0.1.4
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
protobuf==4.23.0
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.17
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-fid==0.3.0
pytorch-lightning==2.0.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rotary-embedding-torch==0.2.3
rsa==4.9
safetensors==0.3.1
scikit-image==0.16.2
scipy==1.10.1
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
timm==0.9.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchmetrics==0.11.4
torchsummary==1.5.1
torchvision==0.13.1+cu113
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
